1-add corner image
sometimes the warning that says that you've already used a customer PO number blocks the corner image we are trying to find since when it appears its location is random. added an extra image for identification that is actually located on the popup so that the image search doesn't fail.