2-finish data-driven portion
this script is now driven by an excel spreadsheet that contains CO numbers to be processed. currently it will take each CO value and run it through the picking process. TODO implement manufacturing in M3 and KPack using the doFullRun function then transfer the working code over to the main script where it needs to be testing with an excel spreadsheet that can be populated using the function generateMOCO